+++ ptychodus/api/data.py
@@ -1,8 +1,10 @@
 from abc import ABC, abstractmethod, abstractproperty
 from collections.abc import Sequence
 from dataclasses import dataclass
+from decimal import Decimal
 from enum import Enum, auto
 from pathlib import Path
+from typing import overload, Optional, Union
 
 import numpy
 import numpy.typing
@@ -10,42 +12,54 @@ import numpy.typing
 from .observer import Observable
 from .tree import SimpleTreeNode
 
-DataArrayType = numpy.typing.NDArray[numpy.integer]
+DiffractionArrayType = numpy.typing.NDArray[numpy.integer]
 
 
-class DatasetState(Enum):
-    NOT_FOUND = auto()
-    EXISTS = auto()
-    VALID = auto()
+class DiffractionDataState(Enum):
+    MISSING = auto()
+    FOUND = auto()
+    LOADED = auto()
 
 
-class DiffractionDataset(Sequence[DataArrayType], Observable):
-
-    @abstractproperty
-    def datasetName(self) -> str:
-        pass
+class DiffractionData(Observable):
 
     @abstractproperty
-    def datasetState(self) -> DatasetState:
+    def name(self) -> str:
         pass
 
     @abstractmethod
-    def getArray(self) -> DataArrayType:
+    def getState(self) -> DiffractionDataState:
+        pass
+
+    @abstractmethod
+    def getStartIndex(self) -> int:
+        pass
+
+    @abstractmethod
+    def getArray(self) -> DiffractionArrayType:
         pass
 
 
 @dataclass(frozen=True)
-class DataFileMetadata:
+class DiffractionMetadata:
     filePath: Path
     imageWidth: int
     imageHeight: int
     totalNumberOfImages: int
+    detectorNumberOfPixelsX: Optional[int] = None
+    detectorNumberOfPixelsY: Optional[int] = None
+    detectorPixelSizeXInMeters: Optional[Decimal] = None
+    detectorPixelSizeYInMeters: Optional[Decimal] = None
+    detectorDistanceInMeters: Optional[Decimal] = None
+    cropCenterXInPixels: Optional[int] = None
+    cropCenterYInPixels: Optional[int] = None
+    probeEnergyInElectronVolts: Optional[Decimal] = None
 
 
-class DataFile(Sequence[DiffractionDataset], Observable):
+class DiffractionDataset(Sequence[DiffractionData], Observable):
 
-    @abstractproperty
-    def metadata(self) -> DataFileMetadata:
+    @abstractmethod
+    def getMetadata(self) -> DiffractionMetadata:
         pass
 
     @abstractmethod
@@ -53,7 +67,38 @@ class DataFile(Sequence[DiffractionDataset], Observable):
         pass
 
 
-class DataFileReader(ABC):
+class SimpleDiffractionDataset(DiffractionDataset):
+
+    def __init__(self, metadata: DiffractionMetadata, contentsTree: SimpleTreeNode,
+                 dataList: list[DiffractionData]) -> None:
+        self._metadata = metadata
+        self._contentsTree = contentsTree
+        self._dataList = dataList
+
+    def getMetadata(self) -> DiffractionMetadata:
+        return self._metadata
+
+    def getContentsTree(self) -> SimpleTreeNode:
+        return self._contentsTree
+
+    @overload
+    def __getitem__(self, index: int) -> DiffractionData:
+        ...
+
+    @overload
+    def __getitem__(self, index: slice) -> Sequence[DiffractionData]:
+        ...
+
+    def __getitem__(
+            self, index: Union[int,
+                               slice]) -> Union[DiffractionData, Sequence[DiffractionData]]:
+        return self._dataList[index]
+
+    def __len__(self) -> int:
+        return len(self._dataList)
+
+
+class DiffractionFileReader(ABC):
 
     @abstractproperty
     def simpleName(self) -> str:
@@ -64,11 +109,11 @@ class DataFileReader(ABC):
         pass
 
     @abstractmethod
-    def read(self, filePath: Path) -> DataFile:
+    def read(self, filePath: Path) -> DiffractionDataset:
         pass
 
 
-class DataFileWriter(ABC):
+class DiffractionFileWriter(ABC):
 
     @abstractproperty
     def simpleName(self) -> str:
@@ -79,6 +124,5 @@ class DataFileWriter(ABC):
         pass
 
     @abstractmethod
-    def write(self, filePath: Path, array: DataArrayType) -> None:
-        # TODO this interface should use a DataFile rather than DataArray
+    def write(self, filePath: Path, dataset: DiffractionDataset) -> None:
         pass
