+++ ptychodus/model/object/core.py
@@ -10,7 +10,8 @@ from ...api.object import (ObjectArrayType, ObjectFileReader, ObjectFileWriter,
 from ...api.observer import Observable, Observer
 from ...api.plugins import PluginChooser, PluginEntry
 from ...api.settings import SettingsRegistry
-from ..data import CropSizer, Detector
+from ..data import CropSizer
+from ..detector import Detector
 from ..probe import Apparatus, ProbeSizer
 from ..scan import Scan
 from .file import FileObjectInitializer
