--- ptychodus/plugins/velociprobe/__init__.py
@@ -1,10 +1,0 @@
-from ptychodus.api.plugins import PluginRegistry
-
-from .velociprobeDataFile import VelociprobeDataFileReader
-from .velociprobeScanFile import VelociprobeScanFileReader
-
-
-def registerPlugins(registry: PluginRegistry) -> None:
-    dataFileReader = VelociprobeDataFileReader()
-    registry.registerPlugin(dataFileReader)
-    registry.registerPlugin(VelociprobeScanFileReader(dataFileReader))
